@@ -23,60 +23,62 @@
   },
   "homepage": "https://github.com/hbz/oerworldmap-ui#readme",
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.1",
-    "babel-preset-env": "^1.6.0",
+    "babel-eslint": "^8.0.2",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
-    "css-loader": "^0.28.4",
-    "enzyme": "^2.9.1",
-    "eslint": "^4.3.0",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-config-stylelint": "^7.0.0",
-    "eslint-plugin-import": "^2.7.0",
+    "css-loader": "^0.28.7",
+    "enzyme": "^3.1.1",
+    "enzyme-adapter-react-16": "^1.0.4",
+    "eslint": "^4.10.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-stylelint": "^8.0.0",
+    "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
-    "jsdom": "^11.1.0",
+    "eslint-plugin-react": "^7.4.0",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.5",
+    "jsdom": "^11.3.0",
     "json-loader": "^0.5.7",
-    "mocha": "^3.5.0",
-    "postcss": "^6.0.8",
-    "postcss-loader": "^2.0.6",
-    "react-test-renderer": "^15.6.1",
-    "style-loader": "^0.18.2",
-    "stylelint": "^8.0.0",
+    "mocha": "^4.0.1",
+    "postcss": "^6.0.14",
+    "postcss-loader": "^2.0.8",
+    "react-test-renderer": "^16.0.0",
+    "style-loader": "^0.19.0",
+    "stylelint": "^8.2.0",
     "stylelint-config-standard": "^17.0.0",
     "stylelint-webpack-plugin": "^0.9.0",
-    "webpack-merge": "^4.1.0"
+    "webpack-merge": "^4.1.1"
   },
   "dependencies": {
-    "asset-require-hook": "^1.1.0",
-    "babel-core": "^6.25.0",
+    "asset-require-hook": "^1.2.0",
+    "babel-core": "^6.26.0",
     "dotenv": "^4.0.0",
     "es6-promise": "^4.1.1",
-    "express": "^4.15.3",
+    "express": "^4.16.2",
     "font-awesome": "^4.7.0",
     "getenv": "^0.7.0",
     "http-link-header": "^0.8.0",
     "ignore-styles": "^5.0.1",
     "isomorphic-fetch": "^2.2.1",
+    "json-pointer": "^0.6.0",
     "json-pointer-form": "file:local_modules/json-pointer-form",
-    "mapbox-gl": "^0.39.1",
+    "mapbox-gl": "^0.41.0",
     "mitt": "^1.1.2",
-    "moment": "^2.18.1",
+    "moment": "^2.19.1",
     "node-polyglot": "^2.2.2",
     "normalize.css": "^7.0.0",
     "postcss-cssnext": "^3.0.2",
-    "postcss-import": "^10.0.0",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "postcss-import": "^11.0.0",
+    "prop-types": "^15.6.0",
+    "react": "^16.1.0",
+    "react-dom": "^16.1.0",
     "react-hot-loader": "^3.1.2",
     "react-markdown": "^2.5.0",
     "react-popperjs": "0.0.3",
-    "webpack": "^3.4.1",
+    "webpack": "^3.8.1",
     "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.18.2"
+    "webpack-hot-middleware": "^2.20.0"
   }
 }
