@@ -7,6 +7,7 @@ import translate from './translate'
 import Metadata from './Metadata'
 import { formatURL } from '../common'
 import Link from './Link'
+import ResourceTable from './ResourceTable'
 import withEmitter from './withEmitter'
 
 import '../styles/WebPage.pcss'
@@ -135,40 +136,8 @@ const WebPage = ({
               ))
             }
 
-            {/* Example of data, GENERATE THIS */}
-            <table>
-              <tbody>
-                <tr>
-                  <td>Location</td>
-                  <td>
-                    Whitehurst Freeway<br />
-                    Washington <br />
-                    United States
-                  </td>
-                </tr>
-                <tr>
-                  <td>Tags</td>
-                  <td>
-                    OER
-                  </td>
-                </tr>
-                <tr>
-                  <td>Creator</td>
-                  <td>
-                    Katy Jordan
-                  </td>
-                </tr>
-                <tr>
-                  <td>Entries mentioned</td>
-                  <td>
-                    The Saylor Academy <br />
-                    OER Hub
-                  </td>
-                </tr>
-              </tbody>
-            </table>
+            <ResourceTable value={about} schema={schema} />
 
-            <pre>{JSON.stringify(about, null, 2)}</pre>
           </div>
         )}
       </div>
