@@ -15,8 +15,8 @@ describe('<I18nProvider />', () => {
     for (const language in expected) {
       const wrapper = mount(
         <I18nProvider locales={[language]}><div /></I18nProvider>
-      )
-      assert.equal(wrapper.node.t('PagedCollection.totalItems', {smart_count: 1}), expected[language])
+      ).instance()
+      assert.equal(wrapper.t('PagedCollection.totalItems', {smart_count: 1}), expected[language])
     }
   })
 })
